+++ src/Trarizon.Library/Collections/Helpers/EnumerableQueries/Helpers/.cs
@@ -18,19 +18,6 @@ public static partial class EnumerableQuery
         return false;
     }
 
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static bool TryNext<T>(this IEnumerator<T> enumerator, [MaybeNullWhen(false)] out T current)
-    {
-        if (enumerator.MoveNext()) {
-            current = enumerator.Current;
-            return true;
-        }
-        else {
-            current = default;
-            return false;
-        }
-    }
-
     private static bool IsCheapEmpty<T>(this IEnumerable<T> source)
     {
         return source.TryGetNonEnumeratedCount(out var count) && count == 0;