+++ src/Trarizon.Library/Collections/Helpers/Utilities/QueryUtil.cs
@@ -3,12 +3,15 @@
 namespace Trarizon.Library.Collections.Helpers.Utilities;
 internal static class QueryUtil
 {
-    public static bool IsInOrder<T>(T left, T right, bool descending) where T : IComparisonOperators<T, T, bool>
-        // desc : prev >= curr
-        // asc  : prev <= curr
-        => descending && left >= right || left <= right;
+    public static bool IsInAscOrder<T>(T former, T latter) where T : IComparisonOperators<T, T, bool>
+        => former <= latter;
 
-    public static bool IsInOrder<T>(T left, T right, IComparer<T> comparer, bool descending)
-        => comparer.Compare(left, right) is var res
-        && descending && res >= 0 || res <= 0;
+    public static bool IsInDescOrder<T>(T former, T latter) where T : IComparisonOperators<T, T, bool>
+        => former >= latter;
+
+    public static bool IsInAscOrder<T>(this IComparer<T> comparer, T former, T latter)
+        => comparer.Compare(former, latter) <= 0;
+
+    public static bool IsInDescOrder<T>(this IComparer<T> comparer, T former, T latter)
+        => comparer.Compare(former, latter) >= 0;
 }