+++ src/Trarizon.Library/ExperimentalDiagnosticIds.cs
@@ -1,0 +1,5 @@
+﻿namespace Trarizon.Library;
+internal static class ExperimentalDiagnosticIds
+{
+    public const string EnumerableHelper_Continue = "TRALIB0001";
+}